pyproject.toml
[project]
name = "parallax"
version = "0.2.0"
description = "Minimal agentic creative production CLI — image generation v0"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=1.0.0",
    "fal-client>=0.13.2",
    "google-genai>=0.1.0",
    "httpx>=0.27",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "whisperx>=3.8.5",
]

[project.scripts]
parallax = "parallax.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/parallax"]
# Ensure non-.py assets under the package make it into the wheel so
# importlib.resources.files("parallax") can resolve them after `uv tool
# install`. Specifically: skills/*.md (loaded by narrative-parallax's
# orchestrator agent) and the existing fonts/captions runtime assets.
artifacts = [
  "src/parallax/skills/*.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
